Move blit serialiser to separate file
